Drop the inspector column; fold identity back into detail
The fourth column was crowding the window: it pushed the detail form
content out of frame, the search field bled across the inspector
boundary, and the "Contacts" title was clipped when the inspector
toggled. With three columns the detail view has room to host the
identity header (photo well + name + role), primary email/phone
quick-copy rows, and groups membership directly, so nothing is lost.

- Delete ContactInspectorView and its project reference.
- Fold the photo well, primary email/phone rows (with copy buttons),
  and the Groups membership toggles into ContactDetailView.
- Remove the inspector toggle + binding from ContactListView /
  ContentView and the `contactInspectorVisible` @AppStorage.
- Simplify LayoutMetrics back to a static `windowMinWidth` of
  sidebar + list + detail (no inspector term).

Entire-Checkpoint: c2dc50ed8a91
