+++ ContactManager/Views/ContactDetailView.swift
@@ -2,25 +2,32 @@
 //  ContactDetailView.swift
 //  ContactManager
 //
-//  The detail column's editable form. Text fields bind directly to the
-//  SwiftData model (autosaved by the context); add/remove field actions and
-//  the birthday toggle go through ContactStore. The avatar + photo controls,
-//  quick-copy email/phone, and group-membership toggles live in the
-//  trailing inspector pane.
+//  The detail column: an identity header (photo well + name/role + quick
+//  primary email/phone) followed by an editable form. Text fields bind
+//  directly to the SwiftData model (autosaved by the context); every other
+//  mutation goes through ContactStore.
 //
 
+import AppKit
 import SwiftData
 import SwiftUI
+import UniformTypeIdentifiers
 
 struct ContactDetailView: View {
     @Environment(\.modelContext) private var context
     @Bindable var contact: Contact
+    @Query(sort: \ContactGroup.name) private var allGroups: [ContactGroup]
+    @State private var isImportingPhoto = false
     @State private var errorMessage: String?
 
     private var store: ContactStore { ContactStore(context) }
 
     var body: some View {
         Form {
+            Section { identityHeader }
+
+            quickInfoSection
+
             Section("Name") {
                 TextField("First Name", text: $contact.firstName)
                 TextField("Last Name", text: $contact.lastName)
@@ -49,6 +56,17 @@ struct ContactDetailView: View {
                 }
             }
 
+            Section("Groups") {
+                if allGroups.isEmpty {
+                    Text("No groups yet. Create one in the sidebar.")
+                        .foregroundStyle(.secondary)
+                } else {
+                    ForEach(allGroups) { group in
+                        Toggle(group.displayName, isOn: membership(in: group))
+                    }
+                }
+            }
+
             Section("Notes") {
                 TextField("Notes", text: $contact.notes, axis: .vertical)
                     .lineLimit(3 ... 10)
@@ -65,6 +83,95 @@ struct ContactDetailView: View {
         } message: { message in
             Text(message)
         }
+    }
+
+    // MARK: - Identity header
+
+    private var identityHeader: some View {
+        HStack(spacing: 16) {
+            photoWell
+            VStack(alignment: .leading, spacing: 2) {
+                Text(contact.fullName)
+                    .font(.title2.weight(.semibold))
+                let role = [contact.jobTitle, contact.company]
+                    .filter { !$0.isEmpty }
+                    .joined(separator: " · ")
+                if !role.isEmpty {
+                    Text(role).foregroundStyle(.secondary)
+                }
+            }
+            Spacer()
+        }
+        .padding(.vertical, 4)
+    }
+
+    /// Tappable avatar offering Choose / Remove Photo.
+    private var photoWell: some View {
+        Menu {
+            Button("Choose Photo…") { isImportingPhoto = true }
+            if contact.photoData != nil {
+                Button("Remove Photo", role: .destructive, action: removePhoto)
+            }
+        } label: {
+            AvatarView(contact: contact, size: 72)
+                .overlay(alignment: .bottomTrailing) {
+                    Image(systemName: "camera.circle.fill")
+                        .symbolRenderingMode(.palette)
+                        .foregroundStyle(.white, .blue)
+                        .font(.system(size: 22))
+                        .background(.white, in: Circle())
+                }
+        }
+        .buttonStyle(.plain)
+        .help("Change Photo")
+        .accessibilityLabel("Change Photo")
+        .fileImporter(
+            isPresented: $isImportingPhoto,
+            allowedContentTypes: [.image]
+        ) { result in
+            handleImport(result)
+        }
+    }
+
+    // MARK: - Quick primary info
+
+    @ViewBuilder
+    private var quickInfoSection: some View {
+        if contact.primaryEmail != nil || contact.primaryPhone != nil {
+            Section {
+                if let email = contact.primaryEmail {
+                    quickRow(label: "Email", value: email)
+                }
+                if let phone = contact.primaryPhone {
+                    quickRow(label: "Phone", value: phone)
+                }
+            }
+        }
+    }
+
+    private func quickRow(label: String, value: String) -> some View {
+        HStack {
+            Text(label)
+                .foregroundStyle(.secondary)
+            Spacer()
+            Text(value)
+                .textSelection(.enabled)
+                .lineLimit(1)
+                .truncationMode(.middle)
+            Button {
+                copyToPasteboard(value)
+            } label: {
+                Image(systemName: "doc.on.doc")
+            }
+            .buttonStyle(.borderless)
+            .help("Copy \(label)")
+            .accessibilityLabel("Copy \(label)")
+        }
+    }
+
+    private func copyToPasteboard(_ value: String) {
+        NSPasteboard.general.clearContents()
+        NSPasteboard.general.setString(value, forType: .string)
     }
 
     // MARK: - Repeatable field sections
@@ -88,7 +195,7 @@ struct ContactDetailView: View {
         }
     }
 
-    // MARK: - Actions
+    // MARK: - Field actions
 
     private func addField(kind: FieldKind) {
         do {
@@ -101,6 +208,59 @@ struct ContactDetailView: View {
     private func delete(_ offsets: IndexSet, from fields: [ContactField]) {
         do {
             try store.delete(offsets.map { fields[$0] })
+        } catch {
+            errorMessage = error.localizedDescription
+        }
+    }
+
+    // MARK: - Group membership
+
+    private func membership(in group: ContactGroup) -> Binding<Bool> {
+        Binding(
+            get: { contact.groups.contains { $0.persistentModelID == group.persistentModelID } },
+            set: { isMember in
+                do {
+                    try store.setMembership(of: contact, in: group, isMember: isMember)
+                } catch {
+                    errorMessage = error.localizedDescription
+                }
+            }
+        )
+    }
+
+    // MARK: - Photo actions
+
+    private func handleImport(_ result: Result<URL, Error>) {
+        switch result {
+        case .failure(let error):
+            errorMessage = error.localizedDescription
+        case .success(let url):
+            Task {
+                // Read the file and run the avatar pipeline off the main actor
+                // so a multi-megabyte photo can't hitch the UI.
+                let avatar: Data? = await Task.detached {
+                    let didAccess = url.startAccessingSecurityScopedResource()
+                    defer { if didAccess { url.stopAccessingSecurityScopedResource() } }
+                    guard let fileData = try? Data(contentsOf: url) else { return nil }
+                    return ImageProcessing.avatarData(from: fileData)
+                }.value
+
+                guard let avatar else {
+                    errorMessage = "That file couldn't be read as an image."
+                    return
+                }
+                do {
+                    try store.setPhotoData(avatar, on: contact)
+                } catch {
+                    errorMessage = error.localizedDescription
+                }
+            }
+        }
+    }
+
+    private func removePhoto() {
+        do {
+            try store.setPhotoData(nil, on: contact)
         } catch {
             errorMessage = error.localizedDescription
         }
